refactor(hardhat): split coverage/helpers.ts to match gas-analytics shape
